Moving affect_data out of merc.h
